feat(ui): Admin page for import/export data management
- Create dedicated AdminView.vue with all import/export operations
- Move import/export buttons from FilmsView to Admin page
- Add /admin route to router
- Add subtle Admin footer link in NavBar (not in primary nav)
- Each import/export action includes clear title and one-line description

Closes #300
